requirements.txt
alpaca-py>=0.26.0
anthropic>=0.40.0
pandas>=2.2.0
pandas-ta>=0.3.14b
pydantic>=2.9.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.0
APScheduler>=3.10.0
finnhub-python>=2.4.0
plotly>=5.24.0
datasette>=0.65.0
streamlit>=1.40.0
requests>=2.32.0
pytest>=8.3.0
pytest-mock>=3.14.0
pytz>=2024.1
